agents/llm-cost-pilot/pyproject.toml
[project]
name = "llm-cost-pilot"
version = "0.1.0"
description = "AI agent that optimizes LLM API routing costs using Karpathy's AutoResearch pattern"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
